// src/app/model/sasaran-kerja.model.ts
export class sasaranKerja {
    id: number | undefined; 
    idPYD: number | undefined; 
    idStatus: number | undefined;
    namaStatus: string | undefined;
    idKategoriPenilaian: number | undefined;
    namaKategoriPenilaian: string | undefined;
    idPPP: number | undefined; 
    idPPK: number | undefined; 
    tahunPenilaian: number | undefined;
    tarikhHantar: Date | undefined;
    tarikhSah: Date | undefined;
    createdAt: Date | undefined;
    updateAt: Date | undefined; 
}